+++ javascript/stimulus_reflex.js
@@ -6,6 +6,7 @@ import { dispatchLifecycleEvent } from './lifecycle'
 import { allReflexControllers } from './controllers'
 import { uuidv4, debounce, emitEvent, serializeForm } from './utils'
 import Log from './log'
+import Debug from './debug'
 import {
   attributeValue,
   attributeValues,
@@ -33,9 +34,6 @@ let actionCableSubscriptionActive = false
 // A dictionary of all active Reflex operations, indexed by reflexId
 window.reflexes = {}
 
-// Indicates if we should log calls to stimulate, etc...
-let debugging
-
 // Should Reflex playback be restricted to the tab that called it?
 let isolationMode
 
@@ -54,21 +52,46 @@ const createSubscription = controller => {
       received: data => {
         if (!data.cableReady) return
 
+        let reflexOperations = {}
+
+        for (let name in data.operations) {
+          if (data.operations.hasOwnProperty(name)) {
+            for (let i = data.operations[name].length - 1; i >= 0; i--) {
+              if (
+                data.operations[name][i].stimulusReflex ||
+                (data.operations[name][i].detail &&
+                  data.operations[name][i].detail.stimulusReflex)
+              ) {
+                reflexOperations[name] = reflexOperations[name] || []
+                reflexOperations[name].push(data.operations[name][i])
+                data.operations[name].splice(i, 1)
+              }
+            }
+            if (!data.operations[name].length)
+              Reflect.deleteProperty(data.operations, name)
+          }
+        }
+
         let totalOperations = 0
         let reflexData
 
-        const dispatchEvent = data.operations['dispatchEvent']
-        const morph = data.operations['morph']
-        const innerHtml = data.operations['innerHtml']
+        const dispatchEvent = reflexOperations['dispatchEvent']
+        const morph = reflexOperations['morph']
+        const innerHtml = reflexOperations['innerHtml']
 
         ;[dispatchEvent, morph, innerHtml].forEach(operation => {
-          if (operation && operation.length && operation[0].stimulusReflex) {
+          if (operation && operation.length) {
             const urls = Array.from(
-              new Set(operation.map(m => m.stimulusReflex.url))
+              new Set(
+                operation.map(m =>
+                  m.detail ? m.detail.stimulusReflex.url : m.stimulusReflex.url
+                )
+              )
             )
-            if (urls.length !== 1 || urls[0] !== location.href) return
 
+            if (urls.length !== 1 || urls[0] !== location.href) return
             totalOperations += operation.length
+
             if (!reflexData) {
               reflexData = operation[0].detail
                 ? operation[0].detail.stimulusReflex
@@ -83,15 +106,17 @@ const createSubscription = controller => {
           if (!reflexes[reflexId] && !isolationMode) {
             const element = xPathToElement(reflexData.xpath)
             const controllerElement = xPathToElement(reflexData.cXpath)
-            if (element.reflexController == undefined)
-              element.reflexController = {}
+            element.reflexController = element.reflexController || {}
+            element.reflexData = element.reflexData || {}
+            element.reflexError = element.reflexError || {}
+
             element.reflexController[
               reflexId
             ] = stimulusApplication.getControllerForElementAndIdentifier(
               controllerElement,
               reflexData.reflexController
             )
-            if (element.reflexData == undefined) element.reflexData = {}
+
             element.reflexData[reflexId] = reflexData
             dispatchLifecycleEvent('before', element, reflexId)
             registerReflex(reflexData)
@@ -101,9 +126,12 @@ const createSubscription = controller => {
             reflexes[reflexId].totalOperations = totalOperations
             reflexes[reflexId].pendingOperations = totalOperations
             reflexes[reflexId].completedOperations = 0
-            CableReady.perform(data.operations)
+            CableReady.perform(reflexOperations)
           }
-        } else CableReady.perform(data.operations)
+        }
+
+        // run piggy back operations after stimulus reflex behavior
+        CableReady.perform(data.operations)
       },
       connected: () => {
         actionCableSubscriptionActive = true
@@ -112,7 +140,7 @@ const createSubscription = controller => {
       rejected: () => {
         actionCableSubscriptionActive = false
         emitEvent('stimulus-reflex:rejected')
-        if (debugging) console.warn('Channel subscription was rejected.')
+        if (Debug.enabled) console.warn('Channel subscription was rejected.')
       },
       disconnected: willAttemptReconnect => {
         actionCableSubscriptionActive = false
@@ -156,13 +184,13 @@ const extendStimulusController = controller => {
         element.validity &&
         element.validity.badInput
       ) {
-        if (debugging) console.warn('Reflex aborted: invalid numeric input')
+        if (Debug.enabled) console.warn('Reflex aborted: invalid numeric input')
         return
       }
       const options = {}
       if (
         args[0] &&
-        typeof args[0] == 'object' &&
+        typeof args[0] === 'object' &&
         Object.keys(args[0]).filter(key =>
           [
             'attrs',
@@ -179,7 +207,7 @@ const extendStimulusController = controller => {
       const attrs = options['attrs'] || extractElementAttributes(element)
       const reflexId = options['reflexId'] || uuidv4()
       let selectors = options['selectors'] || getReflexRoots(element)
-      if (typeof selectors == 'string') selectors = [selectors]
+      if (typeof selectors === 'string') selectors = [selectors]
       const resolveLate = options['resolveLate'] || false
       const datasetAttribute = stimulusApplication.schema.reflexDatasetAttribute
       const dataset = extractElementDataset(element, datasetAttribute)
@@ -209,9 +237,11 @@ const extendStimulusController = controller => {
         throw 'The ActionCable channel subscription for StimulusReflex was rejected.'
 
       // lifecycle setup
-      if (element.reflexController == undefined) element.reflexController = {}
+      element.reflexController = element.reflexController || {}
+      element.reflexData = element.reflexData || {}
+      element.reflexError = element.reflexError || {}
+
       element.reflexController[reflexId] = this
-      if (element.reflexData == undefined) element.reflexData = {}
       element.reflexData[reflexId] = data
 
       dispatchLifecycleEvent('before', element, reflexId)
@@ -219,7 +249,7 @@ const extendStimulusController = controller => {
       setTimeout(() => {
         const { params } = element.reflexData[reflexId] || {}
         const formData =
-          options['serializeForm'] == false
+          options['serializeForm'] === false
             ? ''
             : serializeForm(element.closest('form'), { element })
 
@@ -234,7 +264,7 @@ const extendStimulusController = controller => {
 
       const promise = registerReflex(data)
 
-      if (debugging) {
+      if (Debug.enabled) {
         Log.request(
           reflexId,
           target,
@@ -287,7 +317,7 @@ const registerReflex = data => {
 
   promise.reflexId = reflexId
 
-  if (debugging) promise.catch(NOOP)
+  if (Debug.enabled) promise.catch(NOOP)
 
   return promise
 }
@@ -443,7 +473,7 @@ const initialize = (application, initializeOptions = {}) => {
     'stimulus-reflex',
     controller || StimulusReflexController
   )
-  debugging = !!debug
+  Debug.set(!!debug)
 }
 
 if (!document.stimulusReflexInitialized) {
@@ -490,7 +520,7 @@ if (!document.stimulusReflexInitialized) {
 
     reflex.completedOperations++
 
-    if (debugging) Log.success(event)
+    if (Debug.enabled) Log.success(event)
 
     if (reflex.completedOperations < reflex.totalOperations) return
 
@@ -510,18 +540,18 @@ if (!document.stimulusReflexInitialized) {
     const promise = reflexes[reflexId].promise
     const subjects = { error: true, halted: true, nothing: true, success: true }
 
-    if (element && subject == 'error') element.reflexError = body
+    if (element && subject === 'error') element.reflexError[reflexId] = body
 
-    promise[subject == 'error' ? 'reject' : 'resolve']({
+    promise[subject === 'error' ? 'reject' : 'resolve']({
       data: promise.data,
       element,
       event,
       toString: () => body
     })
 
-    reflexes[reflexId].finalStage = subject == 'halted' ? 'halted' : 'after'
+    reflexes[reflexId].finalStage = subject === 'halted' ? 'halted' : 'after'
 
-    if (debugging) Log[subject == 'error' ? 'error' : 'success'](event)
+    if (Debug.enabled) Log[subject === 'error' ? 'error' : 'success'](event)
 
     if (element && subjects[subject])
       dispatchLifecycleEvent(subject, element, reflexId)
@@ -532,9 +562,9 @@ export default {
   initialize,
   register,
   get debug () {
-    return debugging
+    return Debug.value
   },
   set debug (value) {
-    debugging = !!value
+    Debug.set(!!value)
   }
 }