+++ lib/generators/stimulus_reflex/templates/config/initializers/stimulus_reflex.rb
@@ -12,7 +12,7 @@ StimulusReflex.configure do |config|
 
   # Customize server-side Reflex logging format, with optional colorization:
   # Available tokens: session_id, session_id_full, reflex_info, operation, reflex_id, reflex_id_full, mode, selector, operation_counter, connection_id, connection_id_full, timestamp
-  # Available colors: green, yellow, blue, magenta, cyan, white
+  # Available colors: red, green, yellow, blue, magenta, cyan, white
   # You can also use attributes from your ActionCable Connection's identifiers that resolve to valid ActiveRecord models
   # eg. if your connection is `identified_by :current_user` and your User model has an email attribute, you can access r.email (it will display `-` if the user isn't logged in)